Add docs to `PostView`.
