+++ MessageStackView/Constraints/MessageLayout.swift
@@ -74,3 +74,24 @@ public extension MessageLayout {
         ])
     }
 }
+
+// MARK: - UIView + MessageLayout
+
+public extension UIView {
+    
+    /// Layout the `UIView` with a common `MessageLayout` use case: `layout`.
+    ///
+    /// - Note:
+    /// Custom layout is supported, simply add `self` as a subview to  a desired `UIView` and
+    /// constrain it accordingly as you would any other view.
+    func addTo(_ layout: MessageLayout) {
+        // Remove from previous layout tree if exists
+        removeFromSuperview()
+        
+        // Constrain `self`
+        layout.constrain(subview: self)
+        
+        // Trigger a layout cycle
+        layout.view.setNeedsLayout()
+    }
+}
