+++ MessageStackView/Extensions/Post/UIView+PostView.swift
@@ -1,0 +1,14 @@
+//
+//  UIWindow+PostView.swift
+//  MessageStackView
+//
+//  Created by Ben Shutt on 08/07/2020.
+//  Copyright © 2020 Ben Shutt. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIView {
+    
+}
